refactor: converting from axios to apollo datasource
