+++ server/context.ts
@@ -2,6 +2,7 @@ import { PubSub } from "apollo-server-express";
 import { User } from "./db";
 import { Response } from "express";
 import { PoolClient } from "pg";
+import { UnsplashApi } from "./schema/unsplash.api";
 
 // creating context that contains both pubsub & current user types that were applied to apollo server in index.ts
 
@@ -10,4 +11,7 @@ export type MyContext = {
   currentUser: User;
   res: Response;
   db: PoolClient;
+  dataSources: {
+    unsplashApi: UnsplashApi;
+  };
 };
