+++ server/package.json
@@ -21,7 +21,6 @@
     "@graphql-codegen/cli": "^1.3.1",
     "@graphql-codegen/typescript": "^1.3.1",
     "@graphql-codegen/typescript-resolvers": "^1.3.1",
-    "@types/axios": "^0.14.0",
     "@types/bcrypt": "^3.0.0",
     "@types/body-parser": "^1.17.0",
     "@types/cookie": "^0.3.3",
@@ -46,8 +45,8 @@
   },
   "dependencies": {
     "@safe-api/middleware": "^0.0.2",
+    "apollo-datasource-rest": "^0.6.1",
     "apollo-server-express": "^2.6.8",
-    "axios": "^0.19.0",
     "bcrypt": "^3.0.6",
     "body-parser": "^1.19.0",
     "cookie": "^0.4.0",
